+++ examples/src/Examples/Fragments/SpectrumShift.hs
@@ -1,13 +1,14 @@
 {-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -Wno-missing-signatures #-}
 
 -- | Example fragment shader: Spectrum Shift.
 module Examples.Fragments.SpectrumShift (fragmentSpectrumShiftShader) where
 
-import Spirdo.Wesl.Reflection (weslShader)
+import Spirdo.Wesl.Reflection (defaultCompileOptions, imports, spirv, wesl)
 
 fragmentSpectrumShiftShader =
-      [weslShader|
+      $(spirv defaultCompileOptions imports [wesl|
 struct Params {
   time_res: vec4<f32>;
   color: vec4<f32>;
@@ -40,4 +41,4 @@ fn main(@builtin(position) frag_coord: vec4<f32>) -> @location(0) vec4<f32> {
 
   return vec4(col.x * vignette, col.y * vignette, col.z * vignette, 1.0) * params.color;
 }
-|]
+|])