+++ examples/src/Examples/Vertex.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -Wno-missing-signatures #-}
 
 -- | Example vertex shaders.
@@ -7,10 +8,10 @@ module Examples.Vertex
   , vertexFullscreenShader
   ) where
 
-import Spirdo.Wesl.Reflection (weslShader)
+import Spirdo.Wesl.Reflection (defaultCompileOptions, imports, spirv, wesl)
 
 vertexShader =
-      [weslShader|
+      $(spirv defaultCompileOptions imports [wesl|
 struct VsOut {
 @builtin(position) position: vec4<f32>;
 @location(0) uv: vec2<f32>;
@@ -21,9 +22,9 @@ fn main(@location(0) in_pos: vec2<f32>, @location(1) in_uv: vec2<f32>) -> VsOut 
 let pos = vec4(in_pos.x, in_pos.y, 0.0, 1.0);
 return VsOut(pos, in_uv);
 }
-|]
+|])
 vertexFullscreenShader =
-      [weslShader|
+      $(spirv defaultCompileOptions imports [wesl|
 struct VsOut {
 @builtin(position) position: vec4<f32>;
 @location(0) uv: vec2<f32>;
@@ -46,4 +47,4 @@ let p = pos[i];
 let t = uv[i];
 return VsOut(vec4(p.x, p.y, 0.0, 1.0), t);
 }
-|]
+|])